+++ distribution/lib/Standard/Database/0.0.0-dev/src/Connection/Postgres_Connection.enso
@@ -25,6 +25,7 @@ from project.Internal.Postgres.Helpers import get_encoding_name, parse_postgres_
 from project.Internal.Upload.Helpers.Default_Arguments import first_column_name_in_structure
 from project.Internal.Upload.Operations.Create import create_table_implementation
 
+polyglot java import org.enso.database.JDBCDriverTypes
 polyglot java import org.enso.database.JDBCProxy
 
 type Postgres_Connection
@@ -40,7 +41,7 @@ type Postgres_Connection
         - `data_link_setup`: The setup for saving the connection as a data link.
     create : Text -> Vector -> (Text -> Text -> Postgres_Connection) -> Data_Link_Setup -> Postgres_Connection
     create url properties make_new data_link_setup =
-        types_record = JDBCProxy.makeTypeRecord "Postgres"
+        types_record = JDBCDriverTypes.create "Postgres"
         jdbc_connection = JDBC_Connection.create types_record url properties
 
         encoding = parse_postgres_encoding (get_encoding_name jdbc_connection)
