+++ distribution/lib/Standard/DuckDB/0.0.0-dev/src/DuckDB_Connection.enso
@@ -21,6 +21,7 @@ from Standard.Database.Internal.Upload.Operations.Create import create_table_imp
 import project.Internal.DuckDB_Entity_Naming_Properties
 import project.Internal.DuckDB_Type_Mapping.DuckDB_Type_Mapping
 
+polyglot java import org.enso.database.JDBCDriverTypes
 polyglot java import org.enso.database.JDBCProxy
 polyglot java import org.enso.duckdb.DuckDBUtils
 
@@ -37,7 +38,7 @@ type DuckDB_Connection
         - `schema`: The initial schema to connect to.
     create : Text -> Vector -> (Text -> Text -> DuckDB_Connection) -> Text -> DuckDB_Connection
     create url properties make_new schema =
-        types_record = JDBCProxy.makeTypeRecord "DuckDB"
+        types_record = JDBCDriverTypes.create "DuckDB"
         jdbc_connection = JDBC_Connection.create types_record url properties schema=schema
 
         ## For initial work using DuckDB, we will use the Postgres dialect.
