+++ distribution/lib/Standard/Microsoft/0.0.0-dev/src/SQLServer_Connection.enso
@@ -28,6 +28,7 @@ import project.Connection.SQLServer_Query.SQLServer_Query
 import project.Internal.SQLServer_Dialect
 import project.Internal.SQLServer_Type_Mapping.SQLServer_Type_Mapping
 
+polyglot java import org.enso.database.JDBCDriverTypes
 polyglot java import org.enso.database.JDBCProxy
 
 type SQLServer_Connection
@@ -43,7 +44,7 @@ type SQLServer_Connection
         - `data_link_setup`: the setup for saving the connection as a data link.
     create : Text -> Vector -> (Text -> SQLServer_Connection) -> Data_Link_Setup -> SQLServer_Connection
     create url properties make_new data_link_setup =
-        types_record = JDBCProxy.makeTypeRecord "SQLServer"
+        types_record = JDBCDriverTypes.create "SQLServer"
         jdbc_connection = JDBC_Connection.create types_record url properties
 
         jdbc_entity_naming_properties = Entity_Naming_Properties.from_jdbc_connection jdbc_connection is_case_sensitive=True
